requirements.txt
asgiref==3.7.2
attrs==23.1.0
coverage==7.2.7
Django==4.2.3
django-ckeditor==6.6.1
django-coverage==1.2.4
django-coverage-plugin==3.1.0
django-filter==23.2
django-jalali==6.0.1
django-js-asset==2.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.3
inflection==0.5.1
jdatetime==4.1.1
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
Pillow==10.0.0
PyJWT==2.8.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.0
rest-framework-simplejwt==0.0.2
rpds-py==0.9.2
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
